Fix Wconversion at distributed_ls

+++ src/distributed_ls/Euclid/Hash_dh.c
@@ -75,12 +75,12 @@ void Hash_dhInit_private(Hash_dh h, HYPRE_Int s)
   h->size = size;
 
 /*
-  hypre_sprintf(msgBuf_dh, "requested size = %i; allocated size = %i", s, size); 
+  hypre_sprintf(msgBuf_dh, "requested size = %i; allocated size = %i", s, size);
   SET_INFO(msgBuf_dh);
 */
 
   /* allocate and zero the hash table */
-  data = h->data = (HashRecord*)MALLOC_DH(size*sizeof(HashRecord)); CHECK_V_ERROR;
+  data = h->data = (HashRecord*)MALLOC_DH((size_t) size * sizeof(HashRecord)); CHECK_V_ERROR;
   for (i=0; i<size; ++i) {
     data[i].key = -1;
     data[i].mark = -1;
@@ -113,13 +113,13 @@ HashData * Hash_dhLookup(Hash_dh h, HYPRE_Int key)
         retval = &(data[idx].data);
         break;
       }
-    } 
+    }
   }
   END_FUNC_VAL(retval)
 }
 
 
-/* 
+/*
   TODO: (1) check for already-inserted  (done?)
         (2) rehash, if table grows too large
 */
